restructured code into separate files moved heap allocated keyring/etc onto stack allocated object updated tests to reflect changes
