+++ strongbox.go
@@ -1,11 +1,8 @@
 package main
 
 import (
-	"bytes"
-	"compress/gzip"
 	"crypto/rand"
 	"crypto/sha256"
-	"encoding/base64"
 	"fmt"
 	"io"
 	"io/ioutil"
@@ -14,63 +11,50 @@ import (
 	"os/exec"
 	"os/user"
 	"path/filepath"
-	"strings"
 
-	"github.com/jacobsa/crypto/siv"
 	"github.com/pkg/errors"
 	"github.com/urfave/cli"
-	"gopkg.in/yaml.v2"
 )
 
 const version = "0.1.0-dev"
 
 var (
-	keyLoader = key
-
-	kr            keyRing
-	prefix        = []byte("# STRONGBOX ENCRYPTED RESOURCE ;")
-	defaultPrefix = []byte("# STRONGBOX ENCRYPTED RESOURCE ; See https://github.com/uw-labs/strongbox\n")
-
 	errKeyNotFound = errors.New("key not found")
+	prefix         = []byte("# STRONGBOX ENCRYPTED RESOURCE ;")
+	defaultPrefix  = []byte("# STRONGBOX ENCRYPTED RESOURCE ; See https://github.com/uw-labs/strongbox\n")
 )
+
+// Strongbox stores application state
+type Strongbox struct {
+	keyring   keyRing
+	keyLoader func(string, keyRing) ([]byte, error)
+}
 
 func main() {
 	log.SetPrefix("strongbox: ")
 	log.SetFlags(log.LstdFlags | log.Lshortfile)
 
 	app := cli.NewApp()
-
 	app.Name = "strongbox"
 	app.Version = version
 	app.Usage = ""
 	app.Description = "Encryption for git users."
 
-	// Set up keyring file name
-	var home string
-	u, err := user.Current()
-	if err != nil {
-		// Possibly compiled without CGO and syscall isn't implemented,
-		// try to grab the environment variable
-		home = os.Getenv("HOME")
-		if home == "" {
-			log.Fatal("Could not call os/user.Current() or find $HOME. Please recompile with CGO enabled or set $HOME")
-		}
-	} else {
-		home = u.HomeDir
+	sb := Strongbox{
+		keyring:   &fileKeyRing{fileName: filepath.Join(getHome(), ".strongbox_keyring")},
+		keyLoader: loadKey,
 	}
-
-	kr = &fileKeyRing{fileName: filepath.Join(home, ".strongbox_keyring")}
 
 	app.Commands = []cli.Command{
 		{
 			Name:        "git-config",
 			Description: "Configure git for strongbox use",
-			Action:      commandGitConfig,
+			Action:      sb.commandGitConfig,
 		},
 		{
 			Name:        "gen-key",
 			Description: "Generate a new key and add it to your strongbox keyring",
-			Action:      commandGenKey,
+			Action:      sb.commandGenKey,
 		},
 		{
 			Name:        "decrypt",
@@ -81,23 +65,23 @@ func main() {
 					Usage: "Private key",
 				},
 			},
-			Action: commandDecrypt,
+			Action: sb.commandDecrypt,
 		},
 
 		{
 			Name:        "clean",
 			Description: "intended to be called internally by git",
-			Action:      commandClean,
+			Action:      sb.commandClean,
 		},
 		{
 			Name:        "smudge",
 			Description: "intended to be called internally by git",
-			Action:      commandSmudge,
+			Action:      sb.commandSmudge,
 		},
 		{
 			Name:        "diff",
 			Description: "intended to be called internally by git",
-			Action:      commandDiff,
+			Action:      sb.commandDiff,
 		},
 
 		{
@@ -110,13 +94,13 @@ func main() {
 		},
 	}
 
-	err = app.Run(os.Args)
+	err := app.Run(os.Args)
 	if err != nil {
 		log.Fatal(err)
 	}
 }
 
-func commandGitConfig(c *cli.Context) (err error) {
+func (sb *Strongbox) commandGitConfig(c *cli.Context) (err error) {
 	args := [][]string{
 		{"config", "--global", "--replace-all", "filter.strongbox.clean", "strongbox clean %f"},
 		{"config", "--global", "--replace-all", "filter.strongbox.smudge", "strongbox smudge %f"},
@@ -135,7 +119,7 @@ func commandGitConfig(c *cli.Context) (err error) {
 	return
 }
 
-func commandDecrypt(c *cli.Context) (err error) {
+func (sb *Strongbox) commandDecrypt(c *cli.Context) (err error) {
 	if !c.IsSet("key") {
 		return errors.New("decrypt requires --key to be set")
 	}
@@ -171,8 +155,8 @@ func commandDecrypt(c *cli.Context) (err error) {
 	return
 }
 
-func commandGenKey(c *cli.Context) (err error) {
-	err = kr.Load()
+func (sb *Strongbox) commandGenKey(c *cli.Context) (err error) {
+	err = sb.keyring.Load()
 	if err != nil && !os.IsNotExist(err) {
 		return
 	}
@@ -185,9 +169,9 @@ func commandGenKey(c *cli.Context) (err error) {
 
 	keyID := sha256.Sum256(key)
 
-	kr.AddKey(c.Args().First(), keyID[:], key)
+	sb.keyring.AddKey(c.Args().First(), keyID[:], key)
 
-	err = kr.Save()
+	err = sb.keyring.Save()
 	if err != nil {
 		return
 	}
@@ -195,311 +179,29 @@ func commandGenKey(c *cli.Context) (err error) {
 	return
 }
 
-func commandClean(c *cli.Context) (err error) {
-	return clean(os.Stdin, os.Stdout, c.Args().First())
+func (sb *Strongbox) commandClean(c *cli.Context) (err error) {
+	return clean(os.Stdin, os.Stdout, c.Args().First(), sb.keyring)
 }
 
-func commandSmudge(c *cli.Context) (err error) {
-	return smudge(os.Stdin, os.Stdout, c.Args().First())
+func (sb *Strongbox) commandSmudge(c *cli.Context) (err error) {
+	return smudge(os.Stdin, os.Stdout, c.Args().First(), sb.keyring)
 }
 
-func commandDiff(c *cli.Context) (err error) {
+func (sb *Strongbox) commandDiff(c *cli.Context) (err error) {
 	return diff(c.Args().First())
 }
 
-func clean(r io.Reader, w io.Writer, filename string) (err error) {
-	in, err := ioutil.ReadAll(r)
+func getHome() (home string) {
+	u, err := user.Current()
 	if err != nil {
-		return
-	}
-
-	// Check the file is plaintext, if its an encrypted strongbox file, copy as is, and exit 0
-	if bytes.HasPrefix(in, prefix) {
-		_, err = io.Copy(w, bytes.NewReader(in))
-		if err != nil {
-			return
+		// Possibly compiled without CGO and syscall isn't implemented,
+		// try to grab the environment variable
+		home = os.Getenv("HOME")
+		if home == "" {
+			log.Fatal("Could not call os/user.Current() or find $HOME. Please recompile with CGO enabled or set $HOME")
 		}
-		return
+	} else {
+		home = u.HomeDir
 	}
-
-	// File is plaintext and needs to be encrypted, get the key, fail on error
-	key, err := keyLoader(filename)
-	if err != nil {
-		return
-	}
-
-	out, err := encrypt(in, key)
-	if err != nil {
-		return
-	}
-
-	_, err = io.Copy(w, bytes.NewReader(out))
-	if err != nil {
-		return
-	}
-
-	return
-}
-
-func smudge(r io.Reader, w io.Writer, filename string) (err error) {
-	in, err := ioutil.ReadAll(r)
-	if err != nil {
-		return errors.Wrap(err, "failed to read input stream")
-	}
-
-	// file is a non-strongbox file, copy as is and exit
-	if !bytes.HasPrefix(in, prefix) {
-		_, err = io.Copy(w, bytes.NewReader(in))
-		if err != nil {
-			return errors.Wrap(err, "failed to copy to output stream")
-		}
-		return
-	}
-
-	key, err := keyLoader(filename)
-	if err != nil {
-		// don't log error if its keyNotFound
-		switch err {
-		case errKeyNotFound:
-		default:
-			log.Println(err)
-		}
-		// Couldn't load the key, just copy as is and return
-		if _, err = io.Copy(w, bytes.NewReader(in)); err != nil {
-			log.Println(err)
-		}
-		return
-	}
-
-	out, err := decrypt(in, key)
-	if err != nil {
-		log.Println(err)
-		out = in
-	}
-	if _, err = io.Copy(w, bytes.NewReader(out)); err != nil {
-		return
-	}
-
-	return
-}
-
-func diff(filename string) (err error) {
-	f, err := os.Open(filename)
-	if err != nil {
-		return
-	}
-	defer func() {
-		if err = f.Close(); err != nil {
-			return
-		}
-	}()
-
-	_, err = io.Copy(os.Stdout, f)
-	if err != nil {
-		return
-	}
-
-	return
-}
-
-func encrypt(b []byte, key []byte) ([]byte, error) {
-	b = compress(b)
-	out, err := siv.Encrypt(nil, key, b, nil)
-	if err != nil {
-		return nil, err
-	}
-	var buf []byte
-	buf = append(buf, defaultPrefix...)
-	b64 := encode(out)
-	for len(b64) > 0 {
-		l := 76
-		if len(b64) < 76 {
-			l = len(b64)
-		}
-		buf = append(buf, b64[0:l]...)
-		buf = append(buf, '\n')
-		b64 = b64[l:]
-	}
-	return buf, nil
-}
-
-func compress(b []byte) []byte {
-	var buf bytes.Buffer
-	zw := gzip.NewWriter(&buf)
-	_, err := zw.Write(b)
-	if err != nil {
-		log.Fatal(err)
-	}
-	if err := zw.Close(); err != nil {
-		log.Fatal(err)
-	}
-	return buf.Bytes()
-}
-
-func decompress(b []byte) []byte {
-	zr, err := gzip.NewReader(bytes.NewReader(b))
-	if err != nil {
-		log.Fatal(err)
-	}
-	b, err = ioutil.ReadAll(zr)
-	if err != nil {
-		log.Fatal(err)
-	}
-	if err := zr.Close(); err != nil {
-		log.Fatal(err)
-	}
-	return b
-}
-
-func encode(decoded []byte) []byte {
-	b64 := make([]byte, base64.StdEncoding.EncodedLen(len(decoded)))
-	base64.StdEncoding.Encode(b64, decoded)
-	return b64
-}
-
-func decode(encoded []byte) ([]byte, error) {
-	decoded := make([]byte, len(encoded))
-	i, err := base64.StdEncoding.Decode(decoded, encoded)
-	if err != nil {
-		return nil, err
-	}
-	return decoded[0:i], nil
-}
-
-func decrypt(enc []byte, priv []byte) ([]byte, error) {
-	// strip prefix and any comment up to end of line
-	spl := bytes.SplitN(enc, []byte("\n"), 2)
-	if len(spl) != 2 {
-		return nil, errors.New("Couldn't split on end of line")
-	}
-	b64encoded := spl[1]
-	b64decoded, err := decode(b64encoded)
-	if err != nil {
-		return nil, err
-	}
-	decrypted, err := siv.Decrypt(priv, b64decoded, nil)
-	if err != nil {
-		return nil, err
-	}
-	decrypted = decompress(decrypted)
-	return decrypted, nil
-}
-
-func key(filename string) ([]byte, error) {
-	keyID, err := findKey(filename)
-	if err != nil {
-		return []byte{}, err
-	}
-
-	err = kr.Load()
-	if err != nil {
-		return []byte{}, err
-	}
-
-	key, err := kr.Key(keyID)
-	if err != nil {
-		return []byte{}, err
-	}
-
-	return key, nil
-}
-
-func findKey(filename string) ([]byte, error) {
-	path := filepath.Dir(filename)
-	for {
-		if fi, err := os.Stat(path); err == nil && fi.IsDir() {
-			keyFilename := filepath.Join(path, ".strongbox-keyid")
-			if keyFile, err := os.Stat(keyFilename); err == nil && !keyFile.IsDir() {
-				return readKey(keyFilename)
-			}
-		}
-		if path == "." {
-			break
-		}
-		path = filepath.Dir(path)
-	}
-	return []byte{}, fmt.Errorf("failed to find key id for file %s", filename)
-}
-
-func readKey(filename string) ([]byte, error) {
-	fp, err := ioutil.ReadFile(filename)
-	if err != nil {
-		return []byte{}, err
-	}
-
-	b64 := strings.TrimSpace(string(fp))
-	b, err := decode([]byte(b64))
-	if err != nil {
-		return []byte{}, err
-	}
-	if len(b) != 32 {
-		return []byte{}, fmt.Errorf("unexpected key length %d", len(b))
-	}
-	return b, nil
-}
-
-type keyRing interface {
-	Load() error
-	Save() error
-	AddKey(name string, keyID []byte, key []byte)
-	Key(keyID []byte) ([]byte, error)
-}
-
-type fileKeyRing struct {
-	fileName   string
-	KeyEntries []keyEntry
-}
-
-type keyEntry struct {
-	Description string `yaml:"description"`
-	KeyID       string `yaml:"key-id"`
-	Key         string `yaml:"key"`
-}
-
-func (kr *fileKeyRing) AddKey(desc string, keyID []byte, key []byte) {
-	kr.KeyEntries = append(kr.KeyEntries, keyEntry{
-		Description: desc,
-		KeyID:       string(encode(keyID[:])),
-		Key:         string(encode(key[:])),
-	})
-
-}
-
-func (kr *fileKeyRing) Key(keyID []byte) ([]byte, error) {
-	b64 := string(encode(keyID[:]))
-
-	for _, ke := range kr.KeyEntries {
-		if ke.KeyID == b64 {
-			dec, err := decode([]byte(ke.Key))
-			if err != nil {
-				return []byte{}, err
-			}
-			if len(dec) != 32 {
-				return []byte{}, fmt.Errorf("unexpected length of key: %d", len(dec))
-			}
-			return dec, nil
-		}
-	}
-
-	return []byte{}, errKeyNotFound
-}
-
-func (kr *fileKeyRing) Load() error {
-
-	bytes, err := ioutil.ReadFile(kr.fileName)
-	if err != nil {
-		return err
-	}
-
-	err = yaml.Unmarshal(bytes, kr)
-	return err
-}
-
-func (kr *fileKeyRing) Save() error {
-	ser, err := yaml.Marshal(kr)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	return ioutil.WriteFile(kr.fileName, ser, 0600)
+	return home
 }
