+++ strongbox_test.go
@@ -12,6 +12,7 @@ import (
 var (
 	priv  []byte
 	plain = []byte("hello world. this is some plain text for testing")
+	sb    Strongbox
 )
 
 func TestMain(m *testing.M) {
@@ -22,12 +23,14 @@ func TestMain(m *testing.M) {
 		panic(err)
 	}
 
-	keyLoader = testKeyLoader
+	sb = Strongbox{
+		keyLoader: testKeyLoader,
+	}
 
 	os.Exit(m.Run())
 }
 
-func testKeyLoader(string) ([]byte, error) {
+func testKeyLoader(filename string, kr keyRing) ([]byte, error) {
 	return priv, nil
 }
 
@@ -35,10 +38,10 @@ func TestMultipleClean(t *testing.T) {
 	assert := assert.New(t)
 
 	var cleaned bytes.Buffer
-	clean(bytes.NewReader(plain), &cleaned, "")
+	clean(bytes.NewReader(plain), &cleaned, "", sb.keyring)
 
 	var doubleCleaned bytes.Buffer
-	clean(bytes.NewReader(cleaned.Bytes()), &doubleCleaned, "")
+	clean(bytes.NewReader(cleaned.Bytes()), &doubleCleaned, "", sb.keyring)
 
 	assert.Equal(cleaned.String(), doubleCleaned.String())
 }
@@ -47,7 +50,7 @@ func TestSmudgeAlreadyPlaintext(t *testing.T) {
 	assert := assert.New(t)
 
 	var smudged bytes.Buffer
-	smudge(bytes.NewReader(plain), &smudged, "")
+	smudge(bytes.NewReader(plain), &smudged, "", sb.keyring)
 
 	assert.Equal(string(plain), smudged.String())
 }
@@ -56,12 +59,12 @@ func TestRoundTrip(t *testing.T) {
 	assert := assert.New(t)
 
 	var cleaned bytes.Buffer
-	clean(bytes.NewReader(plain), &cleaned, "")
+	clean(bytes.NewReader(plain), &cleaned, "", sb.keyring)
 
 	assert.NotEqual(plain, cleaned.Bytes())
 
 	var smudged bytes.Buffer
-	smudge(bytes.NewReader(cleaned.Bytes()), &smudged, "")
+	smudge(bytes.NewReader(cleaned.Bytes()), &smudged, "", sb.keyring)
 
 	assert.Equal(string(plain), smudged.String())
 }
@@ -70,10 +73,10 @@ func TestDeterministic(t *testing.T) {
 	assert := assert.New(t)
 
 	var cleaned1 bytes.Buffer
-	clean(bytes.NewReader(plain), &cleaned1, "")
+	clean(bytes.NewReader(plain), &cleaned1, "", sb.keyring)
 
 	var cleaned2 bytes.Buffer
-	clean(bytes.NewReader(plain), &cleaned2, "")
+	clean(bytes.NewReader(plain), &cleaned2, "", sb.keyring)
 
 	assert.Equal(cleaned1.String(), cleaned2.String())
 }
@@ -81,9 +84,9 @@ func TestDeterministic(t *testing.T) {
 func BenchmarkRoundTripPlain(b *testing.B) {
 	for n := 0; n < b.N; n++ {
 		var cleaned bytes.Buffer
-		clean(bytes.NewReader(plain), &cleaned, "")
+		clean(bytes.NewReader(plain), &cleaned, "", sb.keyring)
 
 		var smudged bytes.Buffer
-		smudge(bytes.NewReader(cleaned.Bytes()), &smudged, "")
+		smudge(bytes.NewReader(cleaned.Bytes()), &smudged, "", sb.keyring)
 	}
 }
